Resolving a problem with undefined node types.
Node types like 'call_expr' and 'pattern_binding_decl' are still not classes, among others. When one of those would appear, all nodes down the hierarchy weren't created as well.

UnknownNode fills that gap, just to show on the output which nodes aren't defined. It also allows the nodes down in the hierarchy to be created and analyzed.